code reorganization for types and std/no_std
This is just some code reorganization, where weare bringing in all types
uner `crate::types`.

Also updated the `use` statements so that it's easier to consider
`std`/`no_std` when we support. Right now it's really the
`std::error::Error` for which we do not have a good substitute. Also,
need to use crates for some `std::collections` and `std::sync`