Moved `register_defaults` to their proper places.
A Layer knows with whom it should register with and not the other way
round, which was the implementation. That is Ethernet was registering
`EtherType`s whereas they should be registered by respective L3s. Moved
those register functions to their right places.

This is less than ideal still for instance inside `packet` we are
calling all the `foo_layer::register_defaults` functions by hand and
this is quite painful. A cleaner API should be something like this. Ask
application to call `packet::register_defaults`, but that in turn gets
called by some `build.rs` like mechanism, where `register_defaults` of
all the packages are collected and then packed together (perhaps in some
order?), so the User does not have to worry about it.