Another improvement in `dns_name_from_u8`
One of the biggest mistake that we were doing in this particular
function was that we were allocating a `Vec` in a function that was
recursively called `labels_from_u8`. This would result in many
'allocations' that were potentially unused and caused a considerable
slow down. A correct approach is let the calle allocate the `Vec` and
pass a mutable reference to it. This has gotten us a pretty good
performance improvement. Anything more now is considerably more involved
sacrificing code readability.