Implemenation of 'packet_to_json' complete
First pass implementation of outputting json for a packet is complete.
Formatting the `layers` array as a struct with keys as
`layer.short_name` and value as `serialize`d value of the layer. Also
using `hex::serde::encode` to serialize `unprocessed`. Ignoring
`unprocessed` and `data` if the respective fields are `empty` or `None`.