Getting started with `build.rs`
For every layer, we were registering it's handling by defining a
function `register_defaults` that would create an entry into the map of
the upper layer. We had to collect all such `register_defaults`
functions, which was likely error-prone if we 'forget' to add that
`register_defaults` inside `layers/mod.rs`. Now all these functions are
collected through `build.rs` by going through all modules.

The current implementation is a basic working one and can be iterated to
make it better.