housekeeping: Documentation and removed re-exports
From the top level we are avoiding all `pub use module::*` type
re-exports. Till such a point we decide what really needs to be
exported, it's perhaps a good idea to be explicit, so that all users
consistently use it.

Changed from `pub` to `pub(crate)` for `dns::register_defaults`.

Minor fixes to `build.rs`