Removed `Derive[(Clone)]` from structures
Strictly speaking, a `Clone` is not required for structs that implement
a Layer. In fact You cannot give a Clone trait bound to a `dyn Layer`.
Thus presence of `Clone` on many structs was confusing and useless.

Note: `Clone` *may be added* to structs wherever it may be useful, but
strictly speaking it's not 'required' and no code should depend on it
being available.