Refactoring and removed clippy warnings
In the decode functions we were wrapping the decoded layer
(as `Box<dyn Layer + Send>`) in a `RefCell` to obtain interior
mutability. It turns out that is not needed, we can simply use a mutable
`Box<_>` itself. This allows us to also get rid of the 'dummy'
`EmptyLayer` which was needed only for the `RefCell`. This slightly
improves the performance of the decode.

Also, fixed the test cases to reflect the correct number of layers
without the removed `EmptyLayer`.