Regression Fix and possible fix for occassional build failures
As we removed the `RefCell` the last layer was not `push`ed to
`packet.layers`. This caused the last layer data to be not added to the
packet. Now properly `push`ing the last layer and also fixed all the
assertions for the # of layers during packet decode.

We are initializing some `lazy_static` maps. Occassionally a build
failure is observed. This may be due to those maps not properly
initialized. So now we try to make sure `lazy_static::initialize` is
called in each `register_defaults` function properly. This should fix
that occassional build failure issue.