`register_defaults` is called only once now.
This was a problem during testing (and in general also as an API), that
if we call `register_defaults` more than once, the subsequent calls
would error (understandably). Now we are making sure that
`register_defaults` is called only once for 'each' layer.

Also now `RegisterError` returns a proper Error about which layer was
trying to register second time.

This should perhaps fix the stray build issue.