Fixed an issue with `register_encap_type`
The `register_encap_type` was checking for existence of `ENCAP_TYPE_ETH`
regardless of the encap type being registered. In a certain sequence,
this could lead to non-registration of some protocol types and or other
demuxes, which would result in a failed test case. Oddly enough this was
only happening while running a github action and not locally (as somehow
the order of registration didn't matter locally for the test cases).
Fixed this properly.

Also when any of the `register_defaults` fails, we will print the exact
error now, this should help us in debugging the failed test cases.

Just in case we see the error again, added a `fails_always_ignore` test
case that prints the current order of `register_defaults`. This could be
further used to debug the problems if any.